version: remove the version-generating-code
The app isn't always in its git context *and* because we use
submodules we can't rely on the line that gets the GIT_SHA:

$(shell git rev-parse HEAD 2> /dev/null || echo "")

because we might be running from the root directory and it's all a bit
of a mess.

Instead, just ask the environment to provide us GIT_SHA, and if it's
not found or not provided, simply use "development" since that would
explain it in most cases.

In other contexts, the CI will take care of it; it might not be
perfect but it's less hairy than having the app do some Makefile
tricks to figure out what version it's in.

Also: drop the "__time__"/build time variable because it's also
annoying to produce (you can't have an env variable be a shell command
in Docker, and GitHub doesn't provide a satisfying equivalent). Also,
I'm not sure how valuable is its contribution to GIT_SHA which tells
you exactly what and when. Happy to be wrong and put it back, but
for now, goodbye.