npm: upgrade jest-dom to silence some scary warnings
> npm WARN deprecated core-js-pure@3.19.0: core-js-pure@<3.23.3 is no
longer maintained and not recommended for usage due to the number of
issues. Because of the V8 engine whims, feature detection in old
core-js versions could cause a slowdown up to 100x even if nothing is
polyfilled. Some versions have web compatibility issues. Please,
upgrade your dependencies to the actual version of core-js-pure.