css: restore Tailwind
Right. So.

It turns out that mysterious index.css file people were working in its
minified version... they obviously don't.

What happens is webpack is ran locally to build all tailwind assets,
and that's minified into the – git-tracked – index.css file, which
avoid doing all of that in build pipelines later.

We still don't want to use Tailwind but there are a bunch of CSS we
need in there (notably for the email template previews which is why
this whole thing started) so keep the tooling alive for that precise
CSS, but eventually it will all be gone, promise.