all: try and bootstrapp off the GDS app instead
GDS does a bit less than CDS and seem to have a slightly leaner
stack. They have less dependencies which are more up to date, and I
tend to prefer the quality of their changelog.

We're going to try and branch off them instead, and it should probably
save us quite a bit of time in the future.