Revert "pip: upgrade to Flask 2.2"
This reverts commit 91481e699c1ad377bc333a2ddc96aacc34c09ed6.

Turns out flask-login has some rather serious API changes, or as its
author points out, that we use unreliable low-level code that relies
on the internal of flask-login, when it shouldn't.

We should really be using FlaskLoginClient to handle our test
authentication but I couldn't really get to the bottom of it when I
attempted to.

In the meantime revert to Flask/Werkzeug 2.1.x and Flask-Login 0.6.1.

[1]: https://github.com/maxcountryman/flask-login/issues/704