tests: don't check on elements without class
The DSFR contrasts with the GOV.UK Design System in the way it picks
up elements without CSS classes and styles them anyway, as opposed to
the explicit approach of GOV.UK (i.e a `p` will get styled by
the DSFR whereas `p.govuk-body` will get styled by GOV.UK).

In that sense the helper to check for unclassed elements is more noise
than signal so disable it entirely.